[#2235] pilorama: Make `Apply` tests likelier to trigger possible data races
Previously the number of concurrent threads was equal to the batch size.
This way we had "good" condition: batches were likely to be triggered
"by size". In this commit we add more concurrent workers to trigger
a situation where a batch is already full, but new items are still
arriving.

Signed-off-by: Evgenii Stratonikov <e.stratonikov@yadro.com>