[#2267] policer: Prevent data loss outside the container
Local storage node can be outside the container of some object and hold
its single replica. In previous implementation `Policer` considered this
replica as redundant. There is a need to change behavior for such cases:
`Policer` must not mark replica as redundant if it did not find any
valid replicas at the stage of bypassing the container.

Make `Policer.processObject` determine the presence of a local node in a
container in the `processNodes` call loop. Do not consider single
existing replica outside the container as redundant and, accordingly, do
not pass it to the redundant replica's callback.

Signed-off-by: Leonard Lyubich <ctulhurider@gmail.com>