[#2267] policer: Hold objects on "offline" nodes
In the absence of a node in the network map, it is impossible to
determine its belonging to the container in the case of a return. To
prevent the potential loss of a single instance of data, offline nodes
must hold objects.

Make `Policer.processObject` to look up for local node in the network
map before considering object replica as redundant. If the node is
outside the network map, the replica is considered meaningful.

Signed-off-by: Leonard Lyubich <ctulhurider@gmail.com>