morph: Change how a domain is checked for presence in NNS
In previous implementation Morph client checked domain presence using
`isAvailable` method of the NeoFS NNS contract. This method requires TLD
to exist in order to check availability of subdomains. If TLD was
missing, client failed availability check. It was inconvenient, because
this outcome was mixed with the impossibility of verifying the fact,
although it was quite consistent with the 'unavailable' state.

Check domain existence by `ownerOf` method with the following logic:
domain exists if and only if it has owner (i.e. `ownerOf` succeeded).

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>