IR: Don't set empty contract addresses in `Viper` by default
Previously Inner Ring app set addresses of all contracts to empty
strings in `Viper` config container by default. This did not allow
detecting the absence of configurations, and the defaults themselves are
invalid. At the same time, the benefit of setting empty addresses in the
config container was not traced.

Do not modify `contracts` section in `defaultConfiguration` function.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>