fstree: Do not fail iteration over just removed files
A directory is read and files are saved to a local variable. The iteration
over such files may lead to a non-existing files reading due to a normal SN
operation cycle and, therefore, may lead to a returning the OS error to a
caller. Skip just removed (or lost) files as the golang std library does in
similar situations:
https://github.com/golang/go/blob/5f1a0320b92a60ee1283522135e00bff540ea115/src/os/dir_unix.go#L128-L133.

Signed-off-by: Pavel Karpy <carpawell@morphbits.io>