IR: Fix launch mode detection on ENV-based configuration
Previously Inner Ring app configured via ENVs incorrectly determined
launch mode: if `NEOFS_IR_MORPH_ENDPOINT_CLIENT_0_ADDRESS` environment
variable was set, local consensus more was selected. The reason is in
the specifics of structuring the configuration by the Viper, due to
which the installation of a nested key does not mean the installation of
a high-level key. Fixing the run mode setting needs to handle the case
of ENVs.

Make `isLocalConsensusMode` function to pre-check setting of indexed
address key: if set, then we definitely deal with RPC mode. It fixes
problem case described above and along the way the corresponding unit
test.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>