workflow: Fix checkout neofs-node action
Currently, the checkout action in the system tests workflow doesn't handle all
possible GitHub events correctly. For example, when a pull request is created
from a fork, the action fails because it can't find the
branch in the base repository.

This commit fix the Checkout action to handle different GitHub
events.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>