morph: Drop priorities
Once a NeoFS node is connected to an RPC node it must stick to this
connection for as long as it can. A node can only use one RPC safely
(because two RPCs can have different state).
Drop priorities, use the order we get in config instead; do not try to
switch to another node ever, network error is the only reason a new client
is instanced.

Signed-off-by: Pavel Karpy <carpawell@morphbits.io>