workflow: Prevent "No space left" error on GitHub-hosted runners
This commit modifies the workflow for system tests to mitigate the issue of
"No space left" on GitHub's runners. Changes include:

1. Modifying the pytest commands to disable the capture of test output.
This reduces the size of allure-reports and conserves disk space.

2. Updating the version of "peaceiris/actions-gh-pages".
The new version is more efficient and uses less disk space.

A separate commit will disable additional debug output in the neofs-testcases
repository to further reduce the size of test output and save disk space.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>