cli: drop support for binary keys, use wallets
Wallets were introduced way back when in 0.27.5 and that's the way keys should
be stored. We've had enough transition period for people to adjust to using
them. Binary keys are dangerous and inconvenient (NEP-6 is common for all of
the Neo tooling, everyone knows how to use it), therefore we shouldn't provide
support for them.

Related to #2136.

Signed-off-by: Roman Khimov <roman@nspcc.ru>