node: Implement internal `PlacementVectors`
In SDK RC8 it is not possible to pass pivot for sorting as `[]byte`, only as
an oid.ID for object placement. It breaks behaviour for container
estimations code that used to pass a custom pivot
(`"load_announcement_EPOCH"` string). Add sorting code from SDK and the same
test for it. Expected to be removed with the next SDK releases adoption.

Signed-off-by: Pavel Karpy <carpawell@morphbits.io>