workflow: Add a deploying Allure reports to NeoFS #2324
https://github.com/nspcc-dev/neofs-testcases/issues/574

Added a workflow for deploy Allure tests result as static
web page to NeoFS.

This new functionality will help to save any test report as an
object in the NeoFS network, we will become independent of
the limitations of the GitHub branches and pages.

It will also be an "advertisement" for our object storage.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>