Introduce tool to migrate objects from Blobovnicza tree to Peapod
Add `cmd/blobovnicza-to-peapod` application which accepts YAML
configuration file of the storage node and, for each configured shard,
overtakes data from Blobovnicza tree to Peapod created in the parent
directory.

The tool is going to be used for phased and safe rejection of the
Blobovnicza trees and the transition to Peapods.

Refs #2453.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>