subscriber: fix off-by-one in awaitHeight()
Block count is literally a block counter, while block height is the height
written in the block. For block zero the height is zero while we have one block
in the chain and therefore the count is one. Persistent store operates with
heights, therefore comparison should be adjusted by one.

Signed-off-by: Roman Khimov <roman@nspcc.ru>