ir: retry mainnet RPC connection if the node is outdated
We can have a set of RPC nodes configured and we will switch them in runtime,
but if the first node is outdated then the IR node will _always_ fail to
start until RPC node catches up. We can have a better behaviour with node
switching to other RPCs in this phase.

Can't be done this easily for the sidechain since in most cases we have an
internal node and it needs to go through the chain completely.

Refs. #2426.

Signed-off-by: Roman Khimov <roman@nspcc.ru>