node: Drop placement code copy
It was added only due to a temporary SDK's interface problem. SDK RC10 has
solved it.
NOTE: generally, the placement for container estimations changed with this
commit: another pivot is used. We do this change cause SDK used to have a
bug related to the placement func and placement changes are expected
after RC9 -> RC10 update anyway.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>