Add parallel test running
Both sanity tests and a full test run are run in parallel.
This will allow to get the result and a link to the report in NeoFS after 30
minutes after creating a PR or a commit in the repo and get the result of all
tests in a few hours.
This will be useful for developers - now they will be able to get test results
before changes are pushed to the master.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>