cli: add container ownership check
The eACL table can only be set by the container owner or his authorized representative. Similar to `container delete`, it makes sense to add an ownership check to `container set-eacl`.

Precede the ACL extensibility check with an owner check: if the `--no-precheck` flag is set, then the caller must be the owner of the container or provide a session token signed by them.

Refs #2436.

Signed-off-by: Ekaterina Pavlova <ekt@morphbits.io>