storagegroup: Verify homomorphic checksums in `CollectMembers`
Previously, `storagegroup.CollectMembers` function could return error
with text `data must be 64-bytes long` if any of the split members were
without homomorphic checksum or with the invalid one. Including the
bypass of the splitting chain, and the error occurred already after at
stage `tz.Concat`. The correct format of the homomorphic checksum is
required to form the group, so it is required to do an in-place check.

Check each element of `IterateSplitLeaves` iterator and break traversal
on any checksum problem.

Refs #2488.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>