sidechain/deploy: Calculate nonce and vub from NeoFS network state
In order to prevent duplication of update transactions, there is a need
to sync nonce and ValidUntilBlock values in a distributed manner. Since
Sidechain is tied to the particular running NeoFS network, and update
procedure is performed during its lifetime, it makes sense to base these
variables to the network state.

For all update transactions, set:
 * nonce to the current NeoFS epoch;
 * ValidUntilBlock to E+100, where E is a height of the Sidechain at
   which the current epoch began.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>