node: Do not stop estimations iterations ever
Iterations over the locally calculated values should not stop because of any
errors. Sending part of the information is better than sending nothing in this
case. Moreover, the `code = 3072 message = container not found` logs were found,
they are definitely not a good reasons to stop the iterations over the remaining
estimations.
Refs #1506.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>