morph/client: drop group contract call scope, fix #2619
It's not the most elegant way to handle the problem since normally each
contract handler should decide on these things by its own (even per-call),
but given the structure we have this is the easiest "universal" set of rules.
It has a downside in that it MUST be synchronized with contracts changes, but
those are not very frequent, so I think it's good enough. Managing group key
is far worse problem than that.

Signed-off-by: Roman Khimov <roman@nspcc.ru>