morph/deploy: Refactor event listening
Previously, `deploy.Blockchain` interface had same event subscription
interface as Neo RPC WebSocket client in Neo Go lib. Although this made
it possible to directly use `rpcclient.WSClient` as an instance, the
interface itself was over-saturated with respect to its use in the
deployment procedure. In particular, event filtering was never done and
events were never demuxed. Also, the previous version of the interface
made it impossible to use a client connected to several endpoints.
Based on these considerations, the interface is simplified. As a result,
it is now possible to provide an interface from a multi-endpoint client.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>