morph/deploy: Abort waiting for a new block when the channel closes
Deploy procedure sometimes waits for a new block from the blockchain.
Previously, waiting could be interrupted by the global context only.
This did not allow waiting to be interrupted when the event stream was
interrupted (which usually means the connection was lost).

Closing the channel of new blocks will now result in a lost connection
error returned from the deployment procedure.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>