replicator: Demux one message with the replicated object between nodes
Previously, replication service of storage nodes encoded single object
into protocol message N times, where N is a number of nodes to accept
new object replica. Based on the fact that if the object is fixed, the
request remains unchanged, it makes sense to prepare the message once
and send it to all nodes.

Closes #2316. Refs #2317.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>