deploy: Change nonce and ValidUntilBlock settings for contract updates
Previously, NeoFS chain auto-update procedure calculated nonce and
ValidUntilBlock of the update transactions from the latest NeoFS epoch.
This could lead to a deadlock if the epochs were not ticking (and this
could have various reasons). To prevent such problems and at the same
time simplify the formation of transactions, it makes sense to base
height parameters on the current blockchain height only.

Set nonce/ValidUntilBlock to lower/upper multiple of 100 relative to
current height.

Fixes #2680.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>