node/client-cache: Log errors that led to a connection drops
It simplifies connection drop inspection cause after the first connection drop,
there are a lot of `client has recently failed, skipping` logs, and it is not
clear what exact reason did not allow the node to communicate with a "problem"
one. It could be dropped in the future, but a connection loss for 30s (default
reconnection interval for now) is not a normal situation in general, and it
causes replication processes, data duplication, etc., so logs about it can be
kept.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>