fschain/deploy: Fix incorrect Alphabet name->contract mapping in NNS
Previously, NeoFS chain's deployment procedure could set incorrect
mapping for the Alphabet contracts: each contract-bound domain
`alphabetX.neofs` could carry address of any other Alphabet contract.
The address of any Alphabet contract is consistent on the network,
because the contract is deployed by the corresponding fixed account.
Thus, architecturally, nothing should interfere with creating a
corresponding mapping. The problem was caused by an error in the code:
if there was no contract record in the NNS, each node pre-calculated the
address as if it itself had deployed the contract.

It should be noted that the problem did not occur with the remaining
contracts, since each of them has a unique state, regardless of the
deployer. However, for stability reasons, we should prevent the
setting of contract's domain records if the node is not trying to deploy
the contract, and the deployer is not assigned globally (as with
Alphabet).

Fixes #2713.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>