morph/event: rework recursion probe
In some cases we may not receive the original (sent by SN) request or receive
it with a huge delay, so we shouldn't be relying on this magic. Instead, we
can prevent recursion by checking the FB signer and then main tx cache is
sufficient to not try handling the transaction again. Even if we do that, it's
not a huge problem, the fallback will be the same and a single main tx will be
constructed anyway.

Notice that while we could be checking for main/fallbacks in the notary pool,
they can legitimatelly be evicted from it as well, so it's not a 100% proof
of handling status, therefore we're not doing it, relying on tx hash cache and
fallback stability.

Signed-off-by: Roman Khimov <roman@nspcc.ru>