object/search: Prevent transmission of empty intermediate messages
Previously, `ObjectService.Search` RPC server could write empty
intermediate messages to the response stream. This did not break
anything, but it resulted in a waste of cluster resources without any
benefit. The case occurred due to the fact that the server sends at
least one response message based on the search result on each of the
container nodes, and empty results were not excluded.

From now, if nothing is found on a single node, the server does not
write a message to the stream.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>