object/search: Do not access the node repeatedly within the same request
Previously, `ObjectService.Search` RPC server could send a request to a
remote container node more than once in the context of processing a
single client request. On the one hand, this makes no sense, since the
state of the remote node is assumed to be unchanged in the context of a
single request. On the other hand, the resource costs for a repeated
request are prohibitively high - signatures, network, local storage.

In practice, repetitions occurred in two cases:
 - storage policy specifics, e.g. with several replica (`REP`) rules
 - when switching processing to previous epochs: some nodes could
   "remain" in the container

This is solved by storing the processed nodes in memory. Incurs much
lower costs than repeated requests.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>