morph: Fix subscription deadlock
Scenario:
0. at least one subscription has been performed
1. another subscription is being done
2. a notification from one of the `0.` point's subs is received

If `2.` happens b/w `0.` and `1.` a deadlock appears since the notification
routing process is locked on the subscription lock while the subscription lock
cannot be unlocked since the subscription RPC cannot be done before the
just-arrived notification is handled (read from the neo-go subscription
channel).

`switchLock` does the same thing to the `routeNotifications`: it ensures that no
routine is doing/will be doing changes with the subscription channels, even
though `subs`'s lock was created for this purpose initially.

Relates #2559.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>