.github/linter: Use GO version from the go.mod file
It allows not updating it manually. Why an actual version is important: linter
tries to compile the code inside and marks any compiler errors as a `typecheck`
"linter" message, so if we use some GO features that are available only in 1.20+
versions, it is impossible to pass the linter job with 1.19 GO installed on the
worker machine.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>