policer: Reuse memory buffers with encoded replication requests
Previously, `Policer` always allocated a new memory buffer for encoding
the request when replicating the next object. Now used buffers go to the
`sync.Pool` which, keeping them from the GC, will allow them to be
reused if possible. A pool from the gRPC lib was taken as the starting
implementation.

This brings #2316 resolution closer: suggestion is using a static
per-worker buffer. For now, the buffers can be freed so as not to
provoke the residence of unused memory as the amount of work subsides.
This approach may be changed later.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>