Revert "[#2260] services/object: Do not assemble object with TTL=1"
This reverts commit 907f427b99768ff010834621578441ef81dc0ff9.

GET requests with TTL=1 are absolutely correct regardless of the need to
assemble the object. This TTL blocks request forwarding - it does not
happen. At the same time, assembling an object on the server spawns
new requests from scratch, which is not forwarding. The original commit
does not have a description of the cause, and it was never discovered.

Fixes #2447.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>