shard: Set metabase presence flag if object is found in the write cache
Previously, `Shard.Get*` methods returned unset meta presence flag
if when object was requested and found in the metabase. The
documentation did not explain why. Callers do not isolate the case of a
cache hit (this is not visible from the return anyway). In total, this
behavior was recognized as incorrect and fixed.

Refs 69e1e6ca20f1979f982f6bbd6c94f70a1ee92a69.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>