metabase: Do not search for the link object in the regular objects bucket
Link objects are stored in a separate bucket for V2 split. If no link object is
found, fallback to the previous behavior or try to find the last object. Also,
do not try to unmarshal empty object payload, since it leads to hard-to-debug
parsing errors in logs, while not having children objects at all is a correct
case.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>