pkg/morph/container: Do not create client with notary status
Provide notary signature status with explicit parameter. Some calls are always
expected to be Alphabet signed (at least for the current contracts), make it
explicit for a code reader, that is more error proof.
This commit also fixes container size estimations writing: they were tried to be
notary signed, but the Alphabet does not handle such requests. This is also the
reason why this kinda refactor does not touch other contracts; general client
improvements can be done with another PR.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>