adm: never use GetContractStateByID(1) to retrieve NNS hash
And use a bit more of NNS wrapper.

Notice that this changes "alphabet X" to "alphabetX" in dump output, but I
consider this to be an improvement (since that's the contract name).

It also deoptimizes some carefully crafted code to extract a number of names in
one call. Likely it's not a performance issue you care about and it's actually
somewhat broken in that it assumes one result per invocation (but you can get
zero or more than one).

Many more things can be improved, but let's start with this.

Refs. #2826.

Signed-off-by: Roman Khimov <roman@nspcc.ru>