ir/config: Fix unenforced `morph.consensus.p2p.peers.min` config default
As declared in NeoFS IR configuration, min required number of peer
connections must default to ⌈2/3N-1⌉. Previously, IR app did not set
default value when `p2p` and/or `p2p.peers` sections were omitted, only
when `p2p.peers` is presented and `min` value is missing. This was
incorrect and also unresponsive to the admin: a forgotten value
setting had the effect of no consensus and no new blocks.

This fixes the behavior: now value is recalculated in any setting
without a field. Note that while explicitly specifying zero is
problematic in general, it does not default to preserve admin intent.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>