node/policy: Cache container policy application results
Application result of container (C) storage policy to the network map
(N) does not change for fixed C and N. Previously, `Search` and
`Replicate` object server handlers always calculated the list of
container nodes from scratch. This resulted in excessive node resource
consumption when there was a dense flow of requests for a small number
of containers per epoch. The obvious solution is to cache the latest
results.

A similar attempt had already been made earlier with
9269ed344db58cc29189dc664ba381b9063617af, but it turned out to be
incorrect and did not change anything. As can be seen from the code, the
cache was checked only if the pointers of the received network map and
the last processed one matched. The latter was never set, so there were
no cache callsю

This adds a caching component for up to 1000 recently requested lists of
container nodes. By increasing the amount of memory retained, the
component will mitigate load spikes on a small number of containers. The
volume limit of 1000 was chosen heuristically as a first approximation.

Tests on the development environment showed a pretty good improvement,
but results on real load tests are yet to be obtained. Based on this,
similar optimization for other layers and queries will be done later.

Refs #2692.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>