pkg/morph/client: always wait for notary deposits
Adopting neo-go new waiting API functionality. It is a little functional change
in the NewEpoch event handling (deposit was sent but not awaited), but it should
not be a disadvantage: a silent sending without waiting can confuse when reading
logs.
`morph/client/Client.Wait` was also dropped as not used.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>