node/grpc: Fix calculation of max recv message size from object limit
Since storage node serves `ObjectService.Replicate` RPC, the gRPC server
must be able to accept the biggest allowed object. Previously, node set
max size of received gRPC messages to sum of payload (`MaxObjectSize`
network setting) and header (16K at the moment) size limits. This was
not entirely correct because replication request message also contains
non-object fields (only signature for now) and protobuf service bytes.
Thus, when the size of an object approached the maximum allowed, it was
possible to overflow the calculated limit and receive service refuse.

This adds 1 KB to the calculated limit. This is a heuristic value that
is larger than the current protocol version's query extra data, while
supporting small extensions in advance. The exact value is meaningless
given the smallness volume degrees.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>