blobstore: add minimalistic Get benchmark
It tries to push data with 100 threads which is close to the default maximum
combining count. Then reads are done with as many threads as needed trying to
read various objects.

Unfortunately, size is not a real object size here, it can't be because one
can push any garbage to the store with Put, but Get decodes input data and we
need something that looks like an object, so size is a payload size here.

Signed-off-by: Roman Khimov <roman@nspcc.ru>