node/object/put: exchange meta signatures during replication
Initial replication requires nodes to sign object's main meta information and
respond with it. Meta information is not sent on wire and treated as a fixed
ordered NEO's map. Signatures are verified, not stored/send anywhere yet.
It follows the recent API extension: https://github.com/nspcc-dev/neofs-api/issues/299.
Further, this extension is planned to have a contract adoption:
https://github.com/nspcc-dev/neofs-contract/issues/413 and https://github.com/nspcc-dev/neofs-contract/issues/414.
Closes #2876.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>