meta: do not require metabase resync for v2 to v3 update
Migration can be done automatically, let admin's life be a better thing. TS
expiration is taken as the current epoch + 5. It is not critical if TS will live
more. In practice, side effects _may_ be seen only the first 5 hours after an
update, and only returned status code _may_ differ.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>