meta: fix `Delete` test
Behavior has changed a few times since the original implementation. Also, the
test was broken a little in 8107c8d1a9cc2635510988828e5bfeec13656b4d.
Parent/child relations were changed and now are treated differently at higher
levels, so this test is not so important (see new `TestDeleteAllChildren`
below, it plays a more important role in metabase testing), but keeping it in
working condition with correct statements is not a bad idea.

Signed-off-by: Pavel Karpy <carpawell@nspcc.ru>