cli: Control issuer redefinition in the session token signing cmd
Previously, CLI `util sign session-token` command unconditionally
overrode issuer field to the execution wallet. This could lead to an
unexpected behavior when user explicitly specified the issuer in the
token and wanted to just attach signature to it.

This changes the default behavior to save the issuer field and adds
`--force-issuer` flag reproducing the old behavior.

Refs #2947. Refs #2487.

Signed-off-by: Leonard Lyubich <leonard@morphbits.io>