cli: fix overriding of default container and object attributes
Add a check for the `Timestamp` and the `Name` container attributes, so there
will be an error when you create a container with the `Timestamp` attribute
without the `--disable-timestamp` flag or use the `Name` attribute, which is not
equal to the `--name` flag.
Add a check for the `Timestamp`, the `FileName` and
the `__NEOFS__EXPIRATION_EPOCH` object attributes, so
there will be an error when you put an object with the `Timestamp`
attribute without the `--disable-timestamp` flag, with the `FileName`
attribute without the `--disable-filename` flag or use
the `__NEOFS__EXPIRATION_EPOCH` attribute, which is not equal to the expiration
epoch from '--expire-at' or '--lifetime' flags.
Move logic with expiration epoch attribute to `parseObjectAttrs` function.

Closes #1339.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>