morph: drop inactive client mode, fix #2797
Reconnect indefinitely. After the initial probe this should be the behavior
for client, we can not work without blockchain data so we must wait for it as
long as needed, not entering any "inactive" state where the node pretends to
be alive, but can't do anything.

This makes client context (including everything belonging to it) an atomically
updated thing, regular calls fetch it when needed and can break if there are no
current connection. Meanwhile notification processing thread can reconnect when
needed and restore functionality.

Signed-off-by: Roman Khimov <roman@nspcc.ru>