metabase: drop almost all of Prm/Res structures
They:
 * are present in some methods and not in others
 * add substantial cognitive overhead for anyone looking
 * require more code to handle trivial things
 * can allocate more and more of microobjects creating GC pressure

Test wrappers in the same package perfectly suggest what everyone wanted to
have in the first place. So this patch is a pure refactoring and doesn't
change any behavior, but it drops almost all structures. "Almost" because
there are cases where we pass a number of things (mostly as a result) and
having some enclosing structure makes some sense there.

Signed-off-by: Roman Khimov <roman@nspcc.ru>