node: fix panic in error cases, fix #2970
internalErr has some buffer to accept multiple errors, but it only processes
one of them, so if we have some cascading failures with multiple components
reporting an error some of them will end up panicking because of closed
channed. It's not very critical since the node is to die anyway, but it's
nothing good either.

This particular case shouldn't really happen after #2996, but leaving this
channel open is better than panic in case some other events lead to multiple
sends to it.

Signed-off-by: Roman Khimov <roman@nspcc.ru>