cmd: add tool to migrate objects from Peapod to FSTree
Add `cmd/peapod-to-fstree` application which accepts YAML
configuration file of the storage node, for each configured shard,
overtakes data from Peapod to FSTree that has already been configured and
updates metabase `StorageId` indexes.

The tool is going to be used for phased and safe rejection of the
Peapod and the transition to FSTree.

Closes #2924.

Signed-off-by: Andrey Butusov <andrey@nspcc.io>